added windows installer and changed file structure
