+++ package.json
@@ -1,36 +1,73 @@
 {
-  "name": "Alinka",
+  "name": "alinka",
   "version": "0.1.0",
-  "description": "Aplication for Supporting Center for Children",
   "private": true,
+  "author": "Code for Poznan",
+  "description": "Aplication for Supporting Center for Children",
   "main": "main.js",
-  "scripts": {
-    "webpack": "webpack-dev-server",
-    "electron": "electron .",
-    "start": "concurrently \"BROWSER=none npm run webpack\" \"wait-on http://localhost:4172 && electron .\""
-  },
-  "author": "Code for Poznan <hello@codeforpoznan.pl>",
-  "devDependencies": {
-    "electron": "^4.1.3"
-  },
-  "dependencies": {
-    "babel-core": "^6.24.0",
-    "babel-loader": "^6.4.1",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-stage-1": "^6.22.0",
-    "concurrently": "^4.1.0",
-    "docx": "^4.7.1",
-    "file-system": "^2.2.2",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "wait-on": "^3.2.0",
-    "webpack": "^2.3.2",
-    "webpack-dev-server": "^2.4.2"
-  },
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/CodeForPoznan/alinka-electron/issues"
   },
-  "homepage": "https://github.com/CodeForPoznan/alinka-electron#readme"
+  "homepage": "https://github.com/CodeForPoznan/alinka-electron#readme",
+  "scripts": {
+    "webpack:dev": "MODE=dev webpack-dev-server --mode development --progress",
+    "electron:dev": "MODE=dev electron .",
+    "start": "concurrently \"BROWSER=none npm run webpack:dev\" \"wait-on http://localhost:9000 && npm run electron:dev\"",
+    "build": "webpack --mode production && electron-builder --windows"
+  },
+  "dependencies": {
+    "path": "^0.12.7",
+    "docx": "^4.7.1",
+    "file-system": "^2.2.2",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3"
+  },
+  "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
+    "concurrently": "^4.1.0",
+    "css-loader": "^2.0.0",
+    "electron": "^3.0.11",
+    "electron-builder": "^20.38.3",
+    "html-webpack-plugin": "^3.2.0",
+    "style-loader": "^0.23.1",
+    "wait-on": "^3.2.0",
+    "webpack": "^4.27.1",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.3.1"
+  },
+  "build": {
+    "productName": "Alinka",
+    "appId": "com.alinka.app",
+    "asar": true,
+    "win": {
+      "target": "NSIS",
+      "icon": "./assets/icon.png"
+    },
+    "nsis": {
+      "allowToChangeInstallationDirectory": true,
+      "oneClick": false
+    },
+    "directories": {
+      "buildResources": "build",
+      "output": "dist"
+    },
+    "files": [
+      "**/*",
+      "build/**/*",
+      "assets/**/*",
+      "!dist/**/*",
+      "!src/**/*",
+      "node_modules/**/*"
+    ]
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
