+++ src/components/Header.css
@@ -1,0 +1,3 @@
+.app{
+  color: red;
+}
