+++ src/main.css
@@ -1,0 +1,17 @@
+/** Force App wrapper to bounds */
+[data-reactroot] {
+	width: 100vw !important;
+	height: 100vh !important;
+	position: relative;
+}
+/** Reset Defaults **/
+* {
+	margin: 0;
+	padding: 0;
+	box-sizing: border-box;
+}
+
+body {
+  width: 100% !important;
+	height: 100% !important;
+}
